+++ web/ui/react-app/src/index.tsx
@@ -1,0 +1,7 @@
+import './globals';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+ReactDOM.render(<App />, document.getElementById('root'));