+++ web/ui/react-app/src/utils/func.ts
@@ -1,0 +1,1 @@
+export const uuidGen = () => '_' + Math.random().toString(36).substr(2, 9);